Update header and footer, refactor home page code for mobile
