+++ components/Footer.jsx
@@ -1,7 +1,7 @@
 const Footer = () => {
     return (
-        <footer className="pl-32 flex items-center justify-between h-24 px-8 bg-primary bg-opacity-50">
-            <p className="text-base">Made with ⚔️ for our fellow raiders.</p>
+        <footer className="flex items-center justify-center h-24 px-8 bg-primary bg-opacity-50">
+            <p>Made with ⚔️ for our fellow raiders</p>
         </footer>
     );
 };
