+++ pages/index.js
@@ -8,6 +8,7 @@ import Header from '../components/Header';
 
 import clicksWorld from '../public/images/clicks_world.png';
 import handHeart from '../public/images/hand_heart.png';
+import mapImage from '../public/images/map.png';
 
 const Home = () => {
     return (
@@ -15,32 +16,32 @@ const Home = () => {
             <Head>
                 <title>Raid Guild Origins</title>
             </Head>
-            <main className="bg-dark-darker">
-                <Header>
-                    <div className="px-32">
-                        <div className="flex lg:flex-row flex-col justify-around content-center pt-24">
-                            <div>
-                                <h1 className="font-display text-5xl w-[36rem]">
-                                    Every Raider Comes From Somewhere
-                                </h1>
-                                <p className="pt-6 leading-7 w-[36rem]">
-                                    Raiders come from all over the globe, with
-                                    different backgrounds and perspectives. What
-                                    were they doing before joining the Guild?
-                                    What has their experience been like so far
-                                    in Web3? What is life really like working in
-                                    DAO? The answers are in our stories.
-                                </p>
-                            </div>
+            <div className="bg-dark-darker">
+                <div className="h-[650px] relative">
+                    <div className=" relative z-10">
+                        <Header />
+                        <div className="flex flex-col px-6">
+                            <h1 className="font-display text-5xl mt-10">
+                                Every Raider Comes From Somewhere
+                            </h1>
+
+                            <p className="pt-6 leading-7">
+                                Raiders come from all over the globe, with
+                                different backgrounds and perspectives. What
+                                were they doing before joining the Guild? What
+                                has their experience been like so far in Web3?
+                                What is life really like working in DAO? The
+                                answers are in our stories.
+                            </p>
                             <div className="flex">
-                                <div className="w-64 relative h-full">
+                                <div className="relative h-full">
                                     <Image
                                         className="object-cover object-center absolute inset-0 z-50"
                                         src={clicksWorld}
                                         alt="Click world"
                                     />
                                 </div>
-                                <div className="w-56 relative h-full">
+                                <div className="relative h-full">
                                     <Image
                                         className="object-cover object-center absolute inset-0"
                                         src={handHeart}
@@ -50,12 +51,24 @@ const Home = () => {
                             </div>
                         </div>
                     </div>
-                </Header>
-                <div className="container -mt-16 mb-10 px-32">
+
+                    {/* Background Image */}
+                    <Image
+                        src={mapImage}
+                        className="object-cover object-bottom absolute inset-0 z-0"
+                        layout="fill"
+                        draggable="false"
+                        alt="Raid map"
+                    />
+                    <div className="absolute inset-0 bg-gradient-to-b from-dark via-primary opacity-80 z-0"></div>
+                </div>
+
+                {/* Cards */}
+                <div className="container mb-10 px-6">
                     <CardList />
                 </div>
                 <Footer />
-            </main>
+            </div>
         </Fragment>
     );
 };
